+++ sympy/functions/special/tests/test_tensor_functions.py
@@ -1,4 +1,4 @@
-from sympy import symbols, Dij, LeviCivita
+from sympy import symbols, Dij, LeviCivita, KroneckerDelta
 
 x, y = symbols('x,y')
 
@@ -21,4 +21,15 @@ def test_levicivita():
     assert LeviCivita(4, 5, 1, 2, 3) == 1
     assert LeviCivita(4, 5, 2, 1, 3) == -1
 
-
+def test_kronecker_delta():
+    i, j, k = symbols('i j k')
+    D = KroneckerDelta
+    assert D(i, i) == 1
+    assert D(i, i + 1) == 0
+    assert D(0, 0) == 1
+    assert D(0, 1) == 0
+    # assert D(i, i + k) == D(0, k)
+    assert D(i + k, i + k) == 1
+    assert D(i + k, i + 1 + k) == 0
+    assert D(i, j).subs(dict(i=1, j=0)) == 0
+    assert D(i, j).subs(dict(i=3, j=3)) == 1
